requirements.txt
# Python Dependencies for Pet Image Classifier
# Install with: pip install -r requirements.txt

torch>=1.7.0
torchvision>=0.8.0
Pillow>=8.0.0
